% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_formual_terms.R
\name{check_formula_terms}
\alias{check_formula_terms}
\title{StepBeta and StepBetaBinomial internal object}
\usage{
check_formula_terms(model)
}
\arguments{
\item{model}{Beta regression model}
}
\value{
It returns the complete formula in a standard form
}
\description{
StepBeta and StepBetaBinomial internal object
}
