% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combination_terms.R
\name{Combination_Terms}
\alias{Combination_Terms}
\title{StepBetaBinomial internal object}
\usage{
Combination_Terms(Terms, interaction = F)
}
\arguments{
\item{Terms}{Variables from the starting model}

\item{interaction}{Parameter to define which part of linear predictor to operate}
}
\value{
The function create alle possible combination of the linear predictor
}
\description{
StepBetaBinomial internal object
}
