% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dispersion_formula_terms.R
\name{dispersion_formula_terms}
\alias{dispersion_formula_terms}
\title{StepBeta internal object}
\usage{
dispersion_formula_terms(object)
}
\arguments{
\item{object}{full model}
}
\value{
The function updates the formula for the dispersion component of the model
}
\description{
StepBeta internal object
}
