% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Keep_formula_terms.R
\name{keep_formula_terms}
\alias{keep_formula_terms}
\title{StepBeta internal object}
\usage{
keep_formula_terms(the_formula, var_name)
}
\arguments{
\item{the_formula}{Formula of Beta Regression model}

\item{var_name}{Names of the variables to keep}
}
\value{
The function updates the formula, it keeps the variables defined by the user
}
\description{
StepBeta internal object
}
