% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Remove_formula_interactions.R
\name{remove_formula_interactions}
\alias{remove_formula_interactions}
\title{StepBeta internal object}
\usage{
remove_formula_interactions(the_formula)
}
\arguments{
\item{the_formula}{Formula of Beta Regression model}
}
\value{
The function returns a reduced form of the formula. It excludes the interactive effects.
}
\description{
StepBeta internal object
}
